// src/enums/notification-type.ts
export enum NotificationType {
  // Campaign related
  CAMPAIGN_CREATED = 'CAMPAIGN_CREATED',
  CAMPAIGN_APPLICATION_RECEIVED = 'CAMPAIGN_APPLICATION_RECEIVED',
  CAMPAIGN_APPLICATION_ACCEPTED = 'CAMPAIGN_APPLICATION_ACCEPTED',
  CAMPAIGN_APPLICATION_REJECTED = 'CAMPAIGN_APPLICATION_REJECTED',
  CAMPAIGN_WORK_SUBMITTED = 'CAMPAIGN_WORK_SUBMITTED',
  CAMPAIGN_WORK_APPROVED = 'CAMPAIGN_WORK_APPROVED',
  CAMPAIGN_WORK_REJECTED = 'CAMPAIGN_WORK_REJECTED',
  CAMPAIGN_DETAILS_CHANGED = 'CAMPAIGN_DETAILS_CHANGED',
  CAMPAIGN_ENDING_SOON = 'CAMPAIGN_ENDING_SOON',
  CAMPAIGN_ENDED = 'CAMPAIGN_ENDED',
  CAMPAIGN_BUDGET_INCREASED = 'CAMPAIGN_BUDGET_INCREASED',
  CAMPAIGN_DEADLINE_EXTENDED = 'CAMPAIGN_DEADLINE_EXTENDED',

  // Payment related
  PAYMENT_RECEIVED = 'PAYMENT_RECEIVED',
  PAYMENT_SENT = 'PAYMENT_SENT',
  PAYMENT_FAILED = 'PAYMENT_FAILED',
  PAYOUT_PROCESSED = 'PAYOUT_PROCESSED',
  STRIPE_ACCOUNT_VERIFIED = 'STRIPE_ACCOUNT_VERIFIED',
  STRIPE_ACCOUNT_ISSUE = 'STRIPE_ACCOUNT_ISSUE',
  WALLET_BALANCE_LOW = 'WALLET_BALANCE_LOW',

  // Messaging related
  NEW_MESSAGE = 'NEW_MESSAGE',
  NEW_CONVERSATION = 'NEW_CONVERSATION',

  // Meeting related
  MEETING_SCHEDULED = 'MEETING_SCHEDULED',
  MEETING_REMINDER = 'MEETING_REMINDER',
  MEETING_CANCELLED = 'MEETING_CANCELLED',
  MEETING_RESCHEDULED = 'MEETING_RESCHEDULED',

  // Account related
  ACCOUNT_VERIFICATION_REQUIRED = 'ACCOUNT_VERIFICATION_REQUIRED',
  ACCOUNT_VERIFIED = 'ACCOUNT_VERIFIED',
  PROFILE_INCOMPLETE = 'PROFILE_INCOMPLETE',

  // System related
  SYSTEM_MAINTENANCE = 'SYSTEM_MAINTENANCE',
  FEATURE_ANNOUNCEMENT = 'FEATURE_ANNOUNCEMENT',
  SECURITY_ALERT = 'SECURITY_ALERT',
}